requirements.txt
Flask==2.0.3
Flask-SQLAlchemy==2.5.1
Flask-WTF==1.0.1
pandas==1.1.5
openpyxl==3.1.2
Werkzeug==2.0.3
greenlet==0.4.17
pytz==2023.3.post1